Badminton Score board
A scoreboard is a large board for publicly displaying the 
score in a game. Most levels of sport from high school 
and above use at least one scoreboard for keeping score, 
measuring time, and displaying statistics. Scoreboards in 
the past used a mechanical clock and numeral cards to 
display the score. When a point was made, a person 
would put the appropriate digits on a hook. Most 
modern scoreboards use electromechanical or electronic 
means of displaying the score. In these, digits are often 
composed of large dot-matrix or seven-segment displays 
made of incandescent bulbs, light-emitting diodes, or 
electromechanical flip segments. An official or neutral 
person will operate the scoreboard, using 
a control panel.